requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
